Updated route_service.py and route_generation.py to handle errors
Added an empty app/__init__.py as the folder "app" wasn't getting recognised as a package properly for me.

In route_service.py, added a new function to check for datatype of metadata before accessing its geometry.

In route_service.py, the metadata appears to be corrupted sometimes when amenity is set to true, seen below:

At line 100: (  "detail": "'NoneType' object has no attribute 'y'") error occurs when object RoutePoint is created.

This error occurs about 1 in every 7 times of a test run with amenity set to True.

Currently, I added a section line to skip the row if row.geometry is None for each row in final_gdf. This is only a temporary fix.