Made current and next points read from metadata
Not ready yet.

Added a new attribute to metadata called "route_points_gdf", but since this changes how metadata is tracked, it will affect other aspects of the code and the final output.

Made changes to the functions update_metadata and handle_backtrack.

So far, the wrong route geometry bug is not fixed, but the code runs as normal. Still, the code might break unexpectedly.